moved specs around
